requirements.txt
streamlit==1.31.1
langchain==0.1.9
langchain-openai==0.0.7
langchain-community==0.0.21
chromadb==0.4.22
python-dotenv==1.0.1
PyPDF2==3.0.1
tiktoken==0.6.0
python-docx==0.8.11
requests==2.31.0
deep-translator==1.11.4
